# app/api/v1/endpoints/auth.py
import secrets
import string
import uuid
from datetime import datetime, timedelta, timezone
from typing import Any, Optional
from fastapi import APIRouter, Depends, HTTPException, Request, status, Body
from fastapi.security import OAuth2PasswordRequestForm
from sqlalchemy.ext.asyncio import AsyncSession
from sqlalchemy import select, delete, update

from app.api import deps
from app.core import security
from app.core import rate_limit
from app.core.config import settings
from app.models.user import User
from app.models.verification import Verification
from app.models.refresh_session import RefreshSession
from app.schemas.user import UserCreate, UserResponse, Token, EmailRequest, OTPVerifyRequest, PasswordResetRequest

router = APIRouter()

# OBJ-001 Gate 1 decisions (dependency_graph.md, 2026-08-21). Hardcoded here
# (not wired to Settings) per tests/README.md's note to qa-engineer's own
# constants -- update both places together if these thresholds ever change.
MAX_OTP_ATTEMPTS = 5
OTP_RESEND_COOLDOWN_SECONDS = 60
FORGOT_PASSWORD_RATE_LIMIT_PER_MINUTE = 5
VERIFY_OTP_RATE_LIMIT_PER_MINUTE = 10
RESET_PASSWORD_RATE_LIMIT_PER_MINUTE = 10

RESET_PASSWORD_PURPOSE = "reset_password"
GENERIC_OTP_SENT_MESSAGE = "If the email exists, an OTP has been sent."
GENERIC_OTP_INVALID_MESSAGE = "Invalid or expired OTP"


def _generate_otp() -> str:
    """CSPRNG-backed 6-digit OTP (audit finding #2: the old stdlib `random`
    module is not cryptographically secure)."""
    return "".join(secrets.choice(string.digits) for _ in range(6))


async def _check_and_consume_otp(db: AsyncSession, email: str, otp: str) -> Verification:
    """Validates `otp` against the live `reset_password` Verification row for
    `email`, sharing ONE failed-attempt budget between /verify-otp and
    /reset-password (audit finding #2 -- design notes section 2).

    - No live (unexpired) row at all -> generic 400, no state change (only
      /forgot-password creates rows; nothing here should become a row-spam
      vector).
    - Live row, wrong code -> increments `attempts`; once it reaches
      MAX_OTP_ATTEMPTS the row is invalidated (expires_at pulled to "now")
      so it stops matching future "live row" lookups too. Same generic 400
      either way -- lockout, expiry, and wrong-code are indistinguishable
      by design (no new oracle).
    - Live row, correct code -> returned to the caller to finish the
      business flow (verify-otp just reports success; reset-password also
      deletes it after use).
    """
    generic_error = HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=GENERIC_OTP_INVALID_MESSAGE)

    result = await db.execute(
        select(Verification).filter(
            Verification.email == email,
            Verification.purpose == RESET_PASSWORD_PURPOSE,
            Verification.expires_at > datetime.now(timezone.utc),
        )
    )
    verification = result.scalars().first()

    if verification is None:
        raise generic_error

    if verification.code != otp:
        verification.attempts += 1
        if verification.attempts >= MAX_OTP_ATTEMPTS:
            verification.expires_at = datetime.now(timezone.utc)
        await db.commit()
        raise generic_error

    return verification


async def _issue_tokens_and_session(
    db: AsyncSession, user: User, family_id: uuid.UUID, jti: uuid.UUID
) -> dict:
    """Mints one access+refresh token pair and persists the backing
    `refresh_sessions` row (OBJ-002, design notes section 1-2). Shared by
    `/auth/login` (fresh family: family_id == jti) and `/auth/refresh`
    (rotation: family_id carried over from the row being superseded, jti
    freshly generated by the caller so it can also stamp the old row's
    `replaced_by`). Does not commit -- the caller controls the transaction
    boundary (login: nothing else to persist; refresh: also revokes the
    superseded row in the same commit).
    """
    issued_at = datetime.now(timezone.utc)
    expires_at = issued_at + timedelta(days=settings.REFRESH_TOKEN_EXPIRE_DAYS)

    db.add(
        RefreshSession(
            id=jti,
            family_id=family_id,
            user_id=user.id,
            issued_at=issued_at,
            expires_at=expires_at,
        )
    )

    access_token = security.create_access_token(user.email, ver=user.token_version)
    refresh_token_str = security.create_refresh_token(
        user.email,
        ver=user.token_version,
        jti=jti,
        expires_delta=expires_at - issued_at,
    )
    return {
        "access_token": access_token,
        "token_type": "bearer",
        "refresh_token": refresh_token_str,
    }


async def _revoke_active_sessions(db: AsyncSession, *filters: Any, now: datetime) -> None:
    """Sets `revoked_at` on every still-active (`revoked_at IS NULL`)
    `refresh_sessions` row matching `filters` -- the shared shape behind
    OBJ-002's three revocation call sites (logout: one row by `id`; reuse
    detection: a whole family by `family_id`; password reset: every row for
    a `user_id`). Does not commit -- caller controls the transaction.
    """
    await db.execute(
        update(RefreshSession)
        .where(RefreshSession.revoked_at.is_(None), *filters)
        .values(revoked_at=now)
    )


@router.get("/me", response_model=UserResponse)
async def read_current_user(
    current_user: User = Depends(deps.get_current_active_user),
) -> Any:
    return current_user


@router.post("/register", response_model=UserResponse, status_code=status.HTTP_201_CREATED)
async def register(
    user_in: UserCreate,
    db: AsyncSession = Depends(deps.get_db)
) -> Any:
    result = await db.execute(select(User).filter(User.email == user_in.email))
    user = result.scalars().first()
    if user:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="The user with this email already exists in the system.",
        )

    user = User(
        email=user_in.email,
        hashed_password=security.get_password_hash(user_in.password),
        is_active=True,
        is_verified=False
    )
    db.add(user)
    await db.commit()
    await db.refresh(user)
    return user


@router.post("/login", response_model=Token)
async def login(
    form_data: OAuth2PasswordRequestForm = Depends(),
    db: AsyncSession = Depends(deps.get_db)
) -> Any:
    result = await db.execute(select(User).filter(User.email == form_data.username))
    user = result.scalars().first()

    if not user or not security.verify_password(form_data.password, user.hashed_password):
        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Incorrect email or password")

    if not user.is_active:
        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Inactive user")

    new_family_id = uuid.uuid4()
    tokens = await _issue_tokens_and_session(db, user, family_id=new_family_id, jti=new_family_id)
    await db.commit()
    return tokens


@router.post("/forgot-password")
async def forgot_password(
    http_request: Request,
    payload: EmailRequest,
    db: AsyncSession = Depends(deps.get_db)
) -> Any:
    await rate_limit.enforce_rate_limit(
        db,
        scope="forgot_password",
        ip=rate_limit.client_ip(http_request),
        email=payload.email,
        limit=FORGOT_PASSWORD_RATE_LIMIT_PER_MINUTE,
    )

    result = await db.execute(select(User).filter(User.email == payload.email))
    user = result.scalars().first()
    if not user:
        return {"msg": GENERIC_OTP_SENT_MESSAGE}

    # Resend cooldown (design notes section 2): if a row for this
    # (email, purpose) was created within the last OTP_RESEND_COOLDOWN_SECONDS
    # -- whether it's still usable or already locked/expired -- don't rotate
    # it. This is deliberate even for an already-locked-out row: it stops an
    # attacker who just burned their attempt budget from immediately
    # requesting a fresh one and resetting it. Same generic 200 either way.
    existing_result = await db.execute(
        select(Verification)
        .filter(
            Verification.email == payload.email,
            Verification.purpose == RESET_PASSWORD_PURPOSE,
        )
        .order_by(Verification.created_at.desc())
    )
    existing = existing_result.scalars().first()
    if existing is not None:
        cooldown_cutoff = datetime.now(timezone.utc) - timedelta(seconds=OTP_RESEND_COOLDOWN_SECONDS)
        if existing.created_at is not None and existing.created_at > cooldown_cutoff:
            return {"msg": GENERIC_OTP_SENT_MESSAGE}

    # Rotate: invalidate any previous OTPs for this email and issue a fresh one.
    await db.execute(
        delete(Verification).where(
            Verification.email == payload.email,
            Verification.purpose == RESET_PASSWORD_PURPOSE
        )
    )

    otp = _generate_otp()
    verification = Verification(
        email=payload.email,
        code=otp,
        purpose=RESET_PASSWORD_PURPOSE,
        expires_at=datetime.now(timezone.utc) + timedelta(minutes=10)
    )
    db.add(verification)
    await db.commit()

    # MOCK EMAIL SENDER
    print(f"============================================")
    print(f" [EMAIL MOCK] To: {payload.email} | OTP: {otp} ")
    print(f"============================================")

    return {"msg": GENERIC_OTP_SENT_MESSAGE}


@router.post("/verify-otp")
async def verify_otp(
    http_request: Request,
    payload: OTPVerifyRequest,
    db: AsyncSession = Depends(deps.get_db)
) -> Any:
    await rate_limit.enforce_rate_limit(
        db,
        scope="verify_otp",
        ip=rate_limit.client_ip(http_request),
        email=payload.email,
        limit=VERIFY_OTP_RATE_LIMIT_PER_MINUTE,
    )

    await _check_and_consume_otp(db, payload.email, payload.otp)

    return {"msg": "OTP verified successfully"}


@router.post("/reset-password")
async def reset_password(
    http_request: Request,
    payload: PasswordResetRequest,
    db: AsyncSession = Depends(deps.get_db)
) -> Any:
    await rate_limit.enforce_rate_limit(
        db,
        scope="reset_password",
        ip=rate_limit.client_ip(http_request),
        email=payload.email,
        limit=RESET_PASSWORD_RATE_LIMIT_PER_MINUTE,
    )

    verification = await _check_and_consume_otp(db, payload.email, payload.otp)

    result_user = await db.execute(select(User).filter(User.email == payload.email))
    user = result_user.scalars().first()

    if not user:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")

    user.hashed_password = security.get_password_hash(payload.new_password)
    # OBJ-002 (design notes section 3, Gate 1 MANDATORY): bump token_version
    # (invalidates every access token, and every refresh token via the `ver`
    # check in /auth/refresh) AND bulk-revoke every still-live
    # refresh_sessions row for this user, in the same transaction -- defense
    # in depth, not relying on `ver` alone to be the only line of defense.
    user.token_version += 1
    await _revoke_active_sessions(
        db, RefreshSession.user_id == user.id, now=datetime.now(timezone.utc)
    )
    await db.delete(verification)
    await db.commit()

    return {"msg": "Password updated successfully"}


def _parse_jti(raw_jti: Optional[str]) -> Optional[uuid.UUID]:
    """Best-effort UUID parse -- a forged/garbage `jti` claim must fall
    through to 'no matching row', not crash the request (design notes
    section 2: 'covers forged/garbage jti')."""
    if raw_jti is None:
        return None
    try:
        return uuid.UUID(str(raw_jti))
    except (ValueError, AttributeError, TypeError):
        return None


@router.post("/refresh", response_model=Token)
async def refresh_token(
    refresh_token: str = Body(..., embed=True),
    db: AsyncSession = Depends(deps.get_db)
) -> Any:
    """OBJ-002 rotation + reuse detection (design notes section 2). Every
    failure branch below raises the SAME generic 401 -- no oracle over
    which specific case (no row / reuse / expired / ver-mismatch) was hit.
    """
    invalid_token_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Invalid or expired refresh token",
    )

    payload = security.decode_refresh_token_claims(refresh_token)
    email = payload["sub"]
    jti = _parse_jti(payload.get("jti"))

    session_row = None
    if jti is not None:
        result = await db.execute(select(RefreshSession).filter(RefreshSession.id == jti))
        session_row = result.scalars().first()

    if session_row is None:
        # Covers: forged/garbage jti, a purged row, AND every pre-OBJ-002
        # legacy token (no jti claim at all).
        raise invalid_token_exception

    now = datetime.now(timezone.utc)

    if session_row.revoked_at is not None:
        # Reuse of an already-rotated token -- revoke the ENTIRE family,
        # not just this row (design notes section 2's core rationale).
        await _revoke_active_sessions(
            db, RefreshSession.family_id == session_row.family_id, now=now
        )
        await db.commit()
        raise invalid_token_exception

    if session_row.expires_at < now:
        # Ordinary expiry -- NOT treated as reuse, no family action.
        raise invalid_token_exception

    result_user = await db.execute(select(User).filter(User.email == email))
    user = result_user.scalars().first()
    if not user or not user.is_active:
        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="User inactive or not found")

    if payload.get("ver") != user.token_version:
        # Password-reset invalidation (design notes section 3).
        raise invalid_token_exception

    new_jti = uuid.uuid4()
    tokens = await _issue_tokens_and_session(
        db, user, family_id=session_row.family_id, jti=new_jti
    )
    # Flush the INSERT of the new row before pointing the old row's
    # replaced_by FK at it -- otherwise SQLAlchemy may order the old row's
    # UPDATE before the new row's INSERT within the same flush, violating
    # the FK constraint.
    await db.flush()
    session_row.revoked_at = now
    session_row.replaced_by = new_jti
    await db.commit()
    return tokens


@router.post("/logout", status_code=status.HTTP_204_NO_CONTENT)
async def logout(
    refresh_token: str = Body(..., embed=True),
    db: AsyncSession = Depends(deps.get_db),
) -> None:
    """OBJ-002 (design notes section 4): always 204 for any well-formed
    request body -- no validity oracle. Revokes exactly the one
    `refresh_sessions` row matching the submitted token's `jti`, if any. An
    invalid signature, expired token, malformed string, or a well-formed
    token of the wrong type (access tokens never carry a `jti`) all fall
    through to a silent no-op: the caller's desired end state (this session
    unusable) already holds in every one of those cases. A request body
    missing the `refresh_token` field entirely is the one case that's a
    genuine schema failure (422), handled by FastAPI/Pydantic before this
    function ever runs.
    """
    jti = _parse_jti(security.extract_jti_if_present(refresh_token))
    if jti is not None:
        await _revoke_active_sessions(
            db, RefreshSession.id == jti, now=datetime.now(timezone.utc)
        )
        await db.commit()
    return None