# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: api-fa-backend — Reusable Hardened Auth Service
  description: >-
    API-First contract for the reusable FastAPI auth starter ("bloque hiperseguro").
    This file is the full source of truth for all HTTP-facing behavior — it documents
    both the endpoints already implemented (register/login/forgot-password/verify-otp/
    reset-password/refresh) and the OBJ-001 remediation surface (JWT type confusion,
    OTP brute-force hardening, rate limiting), plus the OBJ-002 remediation surface
    (session/token lifecycle: `/auth/logout`, refresh-token rotation with reuse
    detection, and `token_version`-based invalidation on password reset — closes
    audit finding #3). Config/startup-only requirements that have no HTTP surface
    (e.g. SECRET_KEY validation) are intentionally NOT modeled here — see
    `docs/api/obj-001-design-notes.md` §4. Session/token persistence design (the
    `refresh_sessions` table, reuse-detection state machine, `token_version` claim)
    is documented in `docs/api/obj-002-design-notes.md`, not repeated here.
  version: "0.3.0-obj-002"
  contact:
    name: solution-architect (PRJ-003)
servers:
  - url: /api/v1
    description: Versioned API root (matches `settings.API_V1_STR`)
...
security: []  # default: unauthenticated. Individual operations opt into `bearerAuth`.
paths:
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Create a new user account
      description: >-
        Unchanged by OBJ-001. Current behavior explicitly enumerates existing emails
        via 400 — this is a known, separately-tracked decision (OBJ-007), not part of
        this objective's scope. Documented here for completeness only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
              example:
                detail: "The user with this email already exists in the system."
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Exchange credentials for an access/refresh token pair
      description: >-
        Timing side-channel on this endpoint is audit finding #5, scoped to OBJ-003 —
        unaffected by this change. Both tokens returned here carry the explicit `type`
        claim from OBJ-001 (`access_token` decodes to `type=access`, `refresh_token`
        decodes to `type=refresh`).

        **Changed by OBJ-002 (finding #3).** Both tokens now also carry a `ver` claim
        set to the user's current `token_version` at issuance time (see
        `obj-002-design-notes.md` §3) — this is what lets a later password reset
        invalidate every token issued before it. The refresh token additionally
        carries a `jti` claim (a fresh UUID) that identifies a new row this call
        inserts into the `refresh_sessions` table, seeding a new token "family" — see
        `obj-002-design-notes.md` §1-2 for the persistence/rotation model that
        `/auth/refresh` and `/auth/logout` build on. Request/response shape is
        unchanged; only JWT payload contents and this new persistent side effect are
        new.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginForm'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '400':
          description: Invalid credentials or inactive user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
              examples:
                bad_credentials:
                  value: { detail: "Incorrect email or password" }
                inactive:
                  value: { detail: "Inactive user" }
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request a password-reset OTP
      description: >-
        **Changed by OBJ-001 (finding #2).** Response shape and the always-generic 200
        message are unchanged (anti-enumeration property preserved). New behavior:

        - IP-based rate limiting (429) — see `RateLimitError`.
        - A per-(email, purpose) issuance cooldown is enforced **silently**: a repeat
          request within the cooldown window for an email that already has a live,
          non-expired OTP still returns the same generic 200 and does **not** rotate
          the existing code or reset its `attempts` counter. No distinct status code
          or message is used for this case — introducing one would recreate the
          existence oracle this endpoint already avoids.
        - OTP generation moves from `random.choices` to `secrets`-based generation
          (implementation detail, not a contract change — see design notes §3).

        **Note (OBJ-002 boundary):** this endpoint only *issues* the OTP. The actual
        session-invalidation effect (bumping `token_version`, revoking active refresh
        sessions) happens on successful `/auth/reset-password` below, not here — an
        issued-but-unused OTP has no session-lifecycle side effect.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequest'
      responses:
        '200':
          description: >-
            Always returned regardless of whether the email exists (anti-enumeration).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                msg: "If the email exists, an OTP has been sent."
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Check an OTP without consuming it
      description: >-
        **Changed by OBJ-001 (finding #2).** Success/failure response shape is
        unchanged. What changes is failure accounting: every failed check (wrong code,
        expired code, or no live code at all) increments the `attempts` counter on the
        active `Verification` row for that `(email, purpose)` pair — the **same**
        counter that `/auth/reset-password` increments on failure, since this endpoint
        is a free oracle against that row (audit finding #2). Once `attempts` reaches
        the configured maximum, the row is invalidated server-side (fail-closed) and
        every subsequent check — correct code or not — returns the identical 400 below
        until a new OTP is requested via `/auth/forgot-password`. There is deliberately
        **no distinct "locked out" response**: reusing the generic 400 prevents this
        endpoint from telling an attacker whether they hit the attempt cap or just
        guessed wrong.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OTPVerifyRequest'
      responses:
        '200':
          description: OTP is valid, unexpired, and not locked out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                msg: "OTP verified successfully"
        '400':
          description: >-
            Invalid, expired, or attempt-locked-out OTP. All three cases are
            indistinguishable by design (see summary above).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
              example:
                detail: "Invalid or expired OTP"
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Consume an OTP to set a new password
      description: >-
        **Changed by OBJ-001 (finding #2).** Same failure-accounting change as
        `/auth/verify-otp` (shared `attempts` counter, identical fail-closed 400,
        no distinct lockout signal). Success path (OTP consumed, `Verification` row
        deleted) is unchanged.

        **Changed by OBJ-002 (finding #3).** On success, in the same transaction as
        the password update: `User.token_version` is incremented by 1, and every
        currently-active (`revoked_at IS NULL`) row in `refresh_sessions` for this
        user is marked revoked. Effect: every access token and every refresh token
        issued before this moment stops working — access tokens fail their `ver`
        check on their next use (no wait for natural expiry beyond that next request),
        and refresh tokens fail both the `ver` check *and* the session-revoked check
        on their next `/auth/refresh` call, whichever runs first. This closes audit
        finding #3's core scenario: a refresh token stolen before a suspicious
        password reset no longer survives the reset. See `obj-002-design-notes.md` §3
        for the full claim-comparison design and §2 for why reset also bulk-revokes
        `refresh_sessions` rather than relying on `token_version` alone. Response
        shape (`msg`) is unchanged — this is a same-transaction side effect, not a
        new field in the 200 body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Password updated; the OTP is consumed (deleted) and cannot be reused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                msg: "Password updated successfully"
        '400':
          description: >-
            Invalid, expired, or attempt-locked-out OTP (indistinguishable — see
            `/auth/verify-otp`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
              example:
                detail: "Invalid or expired OTP"
        '404':
          description: >-
            OTP was valid but the user record no longer exists (deleted between
            `/forgot-password` and `/reset-password`). Unchanged edge case from
            existing code.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
              example:
                detail: "User not found"
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/refresh:
    post:
      operationId: refreshAccessToken
      tags: [auth]
      summary: Exchange a valid refresh token for a new access/refresh token pair
      description: >-
        **Changed by OBJ-001 (finding #1).** `verify_refresh_token` enforces
        `type == "refresh"`. Rejection status is `401` for all token-validity
        failures (bad signature, expired, malformed, or wrong `type`). The "valid
        token, but user inactive/deleted" branch stays `400` (a business-state error,
        not a credential-validation failure).

        **Changed by OBJ-002 (finding #3) — refresh tokens now rotate, with reuse
        detection.** Every successful call:

        1. Looks up the submitted token's `jti` claim in the `refresh_sessions`
           table. No row found (includes any pre-OBJ-002 token, which has no `jti`
           at all) → `401`.
        2. Row found but `revoked_at IS NOT NULL` → **reuse detected**: this exact
           token was already exchanged once before and is being replayed. The
           entire token *family* (every row sharing its `family_id`, i.e. every
           descendant issued by prior rotations of the same original login) is
           revoked, and the call fails `401` — same generic message as any other
           failure here, no distinguishing signal. See `obj-002-design-notes.md` §2
           for the full state machine and the rationale for revoking the whole
           family rather than just this token.
        3. Row found, not revoked, but past its own `expires_at` → `401` (ordinary
           expiry, not treated as reuse — no family-wide revocation).
        4. Row valid → the submitted token's `payload.ver` is compared against the
           user's current `token_version` (loaded from the same `User` row this
           handler already fetches for the inactive/not-found check — no extra
           query). Mismatch → `401` (this is the password-reset invalidation path,
           `obj-002-design-notes.md` §3).
        5. All checks pass → the current session row is marked revoked
           (`replaced_by` set to the new row's id), a **new** `refresh_sessions` row
           is inserted (same `family_id`, new `jti`), and a **new** access token +
           **new** refresh token are issued and returned. The submitted refresh
           token cannot be used again after this call succeeds.

        Unlike the pre-OBJ-002 contract, the `refresh_token` in the response is
        **never** the same string that was submitted — callers must persist the
        rotated token and discard the old one immediately.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: >-
            New access token **and** new (rotated) refresh token issued. The
            submitted refresh token is now revoked and cannot be reused — a repeat
            call with it will be treated as reuse detection (see above).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '400':
          description: Token was valid but the associated user is inactive or no longer exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
              example:
                detail: "User inactive or not found"
        '401':
          description: >-
            Token is malformed, has an invalid signature, is expired, has
            `type != "refresh"` (includes an access token being replayed here), has
            no matching (or an already-revoked) `refresh_sessions` row — including
            reuse-detected replays and pre-OBJ-002 legacy tokens with no `jti` — or
            has a `ver` claim that no longer matches the user's current
            `token_version` (invalidated by an intervening password reset). All
            causes share this single generic message and status code — deliberately
            indistinguishable, consistent with this service's existing anti-oracle
            design for OTP/lockout failures. Server-side structured logging (OBJ-004
            scope) SHOULD still record which specific cause fired, for incident
            response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
              example:
                detail: "Invalid or expired refresh token"
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Return the authenticated user (proposed — see design notes §5)
      description: >-
        **New endpoint proposed as part of OBJ-001, not yet in `auth.py`.** Today
        `deps.get_current_user` has zero real consumers, which means finding #1's
        fix (rejecting a refresh token used as a Bearer access token) has no HTTP
        surface to contract-test end-to-end. This minimal protected endpoint gives
        `qa-engineer` a concrete target for that regression test and gives every
        project forking this template a canonical "protected route" example. Flagged
        for explicit confirmation — see final summary.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: >-
            Missing/malformed `Authorization` header, invalid signature, expired
            token, missing `sub`, unknown user, `type != "access"` (finding #1's
            fail-closed fix — a refresh token presented here is rejected with this
            same generic message), **or (OBJ-002) `ver` claim not equal to the
            user's current `token_version`** — this last case means the access
            token was issued before the user's most recent password reset and is
            treated as invalidated even though it hasn't naturally expired yet. All
            causes share the identical generic message, per this service's existing
            anti-oracle design.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
              example:
                detail: "Could not validate credentials"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Access token only. `app/api/deps.get_current_user` decodes the token and
        rejects it (401) unless the `type` claim is exactly `"access"` — a refresh
        token is fail-closed rejected here, never accepted as a substitute (audit
        finding #1). See `obj-001-design-notes.md` §1 for the full claim schema.
        As of OBJ-002, the same handler also compares the token's `ver` claim
        against the user's current `token_version` (loaded as part of the same DB
        read already required to fetch the `User` row — no extra query) and
        rejects on mismatch; see `obj-002-design-notes.md` §3. Access tokens are
        deliberately **not** checked against any session/revocation table — they
        stay fully stateless for latency, backed only by their short natural
        expiry and this `ver` check.
  responses:
    ValidationError:
      description: Request body failed schema validation (Pydantic)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
...
  schemas:
    UserCreate:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
          description: >-
            Must contain at least one uppercase letter, one digit, and one special
            character (`!@#$%^&*()_+-=[]{}|;':",./<>?`). Enforced by
            `validate_password_strength` (unchanged by OBJ-001).
    UserResponse:
      type: object
      required: [id, email, is_active, is_verified, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        is_active:
          type: boolean
        is_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
      description: Never includes `hashed_password` (audit-confirmed PASS, unchanged).
...
    MessageResponse:
      type: object
      required: [msg]
      properties:
        msg:
          type: string
    HTTPError:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
      description: Standard FastAPI `HTTPException` shape, used for every non-2xx, non-422 error.
...